+++ CHANGES.md
@@ -14,6 +14,7 @@
   unless Boot is being updated.
 - Correctly handle case when `:source-paths` or `:resource-paths` are set to
   the empty set (`#{}`).
+- Correctly set last modified time when copying classpath resource.
 
 #### Added
 
@@ -25,6 +26,7 @@
 - `boot.pod/pod-name` &mdash; get/set the name of a pod
 - `boot.pod/coord->map` &mdash; dependency vector to map helper function
 - `boot.pod/map->coord` &mdash; map to dependency vector helper function
+- `boot.pod/resource-last-modified` &mdash; returns last modified time of a classpath resource
 - `boot.core/checkout-dirs` &mdash; the set of directories containing checkout dependency files
 
 ##### Boot Options